+++ src/box/tuple_compare.cc
@@ -571,8 +571,8 @@ tuple_compare_slowpath(struct tuple *tuple_a, hint_t tuple_a_hint,
 	bool was_null_met = false;
 	struct tuple_format *format_a = tuple_format(tuple_a);
 	struct tuple_format *format_b = tuple_format(tuple_b);
-	const uint32_t *field_map_a = tuple_field_map(tuple_a);
-	const uint32_t *field_map_b = tuple_field_map(tuple_b);
+	const uint8_t *field_map_a = tuple_field_map(tuple_a);
+	const uint8_t *field_map_b = tuple_field_map(tuple_b);
 	struct key_part *end;
 	const char *field_a, *field_b;
 	enum mp_type a_type, b_type;
@@ -703,7 +703,7 @@ tuple_compare_with_key_slowpath(struct tuple *tuple, hint_t tuple_hint,
 	struct key_part *part = key_def->parts;
 	struct tuple_format *format = tuple_format(tuple);
 	const char *tuple_raw = tuple_data(tuple);
-	const uint32_t *field_map = tuple_field_map(tuple);
+	const uint8_t *field_map = tuple_field_map(tuple);
 	enum mp_type a_type, b_type;
 	if (likely(part_count == 1)) {
 		const char *field;
@@ -1386,8 +1386,8 @@ func_index_compare(struct tuple *tuple_a, hint_t tuple_a_hint,
 	const char *tuple_b_raw = tuple_data(tuple_b);
 	struct tuple_format *format_a = tuple_format(tuple_a);
 	struct tuple_format *format_b = tuple_format(tuple_b);
-	const uint32_t *field_map_a = tuple_field_map(tuple_a);
-	const uint32_t *field_map_b = tuple_field_map(tuple_b);
+	const uint8_t *field_map_a = tuple_field_map(tuple_a);
+	const uint8_t *field_map_b = tuple_field_map(tuple_b);
 	const char *field_a, *field_b;
 	for (uint32_t i = key_part_count; i < cmp_def->part_count; i++) {
 		struct key_part *part = &cmp_def->parts[i];
